Abstract draggable drawer component.
