+++ bundles.config.js
@@ -680,8 +680,8 @@ exports.extensionBundles = [
         'amp-story-bookend',
         'amp-story-tooltip',
         'amp-story-consent',
+        'amp-story-draggable-drawer-header',
         'amp-story-hint',
-        'amp-story-page-attachment-header',
         'amp-story-unsupported-browser-layer',
         'amp-story-viewport-warning-layer',
         'amp-story-info-dialog',
