+++ extensions/amp-story/1.0/amp-story-draggable-drawer-header.css
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-.i-amphtml-story-page-attachment-header {
+.i-amphtml-story-draggable-drawer-header {
   display: flex !important;
   position: relative !important;
   height: 40px !important;
@@ -24,7 +24,7 @@
   z-index: 1 !important;
 }
 
-.i-amphtml-story-page-attachment-theme-dark {
+.i-amphtml-story-draggable-drawer-theme-dark {
   background: rgba(32, 33, 37, 1) !important;
 }
