+++ extensions/amp-story/1.0/amp-story-draggable-drawer.css
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-amp-story-page-attachment {
+.amp-story-draggable-drawer-root {
   display: block !important;
   position: absolute !important;
   top: 0 !important;
@@ -28,49 +28,49 @@ amp-story-page-attachment {
   transition: transform 0.25s cubic-bezier(0.4, 0.0, 1, 1) !important;
 }
 
-amp-story-page-attachment.i-amphtml-story-page-attachment-open {
+.amp-story-draggable-drawer-root.i-amphtml-story-draggable-drawer-open {
   transform: translate3d(0, 0, 0) !important;
   transition: transform 0.4s cubic-bezier(0.0, 0.0, 0.2, 1) !important;
 }
 
-.i-amphtml-story-page-attachment {
+.i-amphtml-story-draggable-drawer {
   height: 100% !important;
 }
 
-.i-amphtml-story-page-attachment-container {
+.i-amphtml-story-draggable-drawer-container {
   height: calc(100% - 40px) !important;
   background: rgba(255, 255, 255, 1) !important;
   overflow-y: auto !important;
   transition: background 0.3s cubic-bezier(0.4, 0.0, 0.2, 1) !important;
 }
 
-.i-amphtml-story-page-attachment-theme-dark .i-amphtml-story-page-attachment-container {
+.i-amphtml-story-draggable-drawer-theme-dark .i-amphtml-story-draggable-drawer-container {
   background: rgba(32, 33, 37, 1) !important;
 }
 
-.i-amphtml-story-page-attachment-container[hidden] {
+.i-amphtml-story-draggable-drawer-container[hidden] {
   display: block !important;
   background: rgba(255, 255, 255, 0.92) !important;
 }
 
-.i-amphtml-story-page-attachment-theme-dark .i-amphtml-story-page-attachment-container[hidden] {
+.i-amphtml-story-draggable-drawer-theme-dark .i-amphtml-story-draggable-drawer-container[hidden] {
   background: rgba(32, 33, 37, 0.82) !important;
 }
 
-.i-amphtml-story-page-attachment-content {
+.i-amphtml-story-draggable-drawer-content {
   opacity: 1 !important;
   transition: opacity 0.3s cubic-bezier(0.0, 0.0, 0.2, 1) !important;
 }
 
-.i-amphtml-story-page-attachment-container[hidden] .i-amphtml-story-page-attachment-content {
+.i-amphtml-story-draggable-drawer-container[hidden] .i-amphtml-story-draggable-drawer-content {
   opacity: 0 !important;
 }
 
-.i-amphtml-story-page-attachment-container[hidden] .i-amphtml-story-page-attachment-content > * {
+.i-amphtml-story-draggable-drawer-container[hidden] .i-amphtml-story-draggable-drawer-content > * {
   display: none !important;
 }
 
-[desktop] amp-story-page-attachment {
+[desktop] .amp-story-draggable-drawer-root {
   display: flex !important;
   align-items: center !important;
   justify-content: center !important;
@@ -79,13 +79,13 @@ amp-story-page-attachment.i-amphtml-story-page-attachment-open {
   transition: opacity 0.15s cubic-bezier(0.4, 0.0, 1, 1), transform 0s linear 0.15s !important;
 }
 
-[desktop] amp-story-page-attachment.i-amphtml-story-page-attachment-open {
+[desktop] .amp-story-draggable-drawer-root.i-amphtml-story-draggable-drawer-open {
   opacity: 1 !important;
   transition: opacity 0.3s cubic-bezier(0.0, 0.0, 0.2, 1) !important;
   transform: translate3d(0, 0, 0) !important;
 }
 
-[desktop] .i-amphtml-story-page-attachment {
+[desktop] .i-amphtml-story-draggable-drawer {
   height: 60vh !important;
   width: 800px !important;
   border-radius: 8px !important;
