+++ extensions/amp-story/1.0/amp-story-vertical.css
@@ -43,11 +43,11 @@ amp-story[i-amphtml-vertical].i-amphtml-element amp-story-page.i-amphtml-element
   transform: none !important;
 }
 
-[i-amphtml-vertical] .i-amphtml-story-page-attachment-container[hidden] .i-amphtml-story-page-attachment-content {
+[i-amphtml-vertical] .i-amphtml-story-draggable-drawer-container[hidden] .i-amphtml-story-draggable-drawer-content {
   opacity: 1 !important;
 }
 
-[i-amphtml-vertical] .i-amphtml-story-page-attachment-container[hidden] .i-amphtml-story-page-attachment-content > * {
+[i-amphtml-vertical] .i-amphtml-story-draggable-drawer-container[hidden] .i-amphtml-story-draggable-drawer-content > * {
   display: block !important;
 }
