+++ extensions/amp-story/1.0/amp-story.css
@@ -17,7 +17,7 @@
 @import './amp-story-access.css';
 @import './amp-story-bling-link.css';
 @import './amp-story-desktop-panels.css';
-@import './amp-story-page-attachment.css';
+@import './amp-story-draggable-drawer.css';
 @import './amp-story-templates.css';
 @import './amp-story-user-overridable.css';
 @import './amp-story-vertical.css';
